explicit cast to Uint32 for when putting 64bit into two 32bit words (MS compiler spits out warning)

+++ storage/ndb/tools/restore/consumer_restore.cpp
@@ -842,8 +842,8 @@ BackupRestore::report_data(unsigned backup_id, unsigned node_id)
     data[2]= node_id;
     data[3]= m_dataCount & 0xFFFFFFFF;
     data[4]= 0;
-    data[5]= m_dataBytes & 0xFFFFFFFF;
-    data[6]= (m_dataBytes >> 32) & 0xFFFFFFFF;
+    data[5]= (Uint32)(m_dataBytes & 0xFFFFFFFF);
+    data[6]= (Uint32)((m_dataBytes >> 32) & 0xFFFFFFFF);
     Ndb_internal::send_event_report(m_ndb, data, 7);
   }
   return true;
@@ -860,8 +860,8 @@ BackupRestore::report_log(unsigned backup_id, unsigned node_id)
     data[2]= node_id;
     data[3]= m_logCount & 0xFFFFFFFF;
     data[4]= 0;
-    data[5]= m_logBytes & 0xFFFFFFFF;
-    data[6]= (m_logBytes >> 32) & 0xFFFFFFFF;
+    data[5]= (Uint32)(m_logBytes & 0xFFFFFFFF);
+    data[6]= (Uint32)((m_logBytes >> 32) & 0xFFFFFFFF);
     Ndb_internal::send_event_report(m_ndb, data, 7);
   }
   return true;
